update: 피연산자를 여러 타입으로 받을 수 있도록 제네렉을 사용하여 기능 확장 (ArithmeticCalculator)
