+++ src/main/java/calculator/ArithmeticCalculator.java
@@ -3,25 +3,7 @@ package calculator;
 import java.util.ArrayList;
 import java.util.List;
 
-//Enum 타입을 활용하여 연산자 타입에 대한 정보를 관리하고 이를 사칙연산 계산기 ArithmeticCalculator 클래스에 활용 해봅니다.
- enum OperatorType {
-    ADD('+'),
-    SUBTRACT('-'),
-    MULTIPLY('*'),
-    DIVIDE('/');
-
-    private final char symbol;
-
-    OperatorType(char symbol) {  // 생성자
-        this.symbol = symbol;
-    }
-
-    public char getSymbol() {
-        return symbol;
-    }
-}
-
-public class ArithmeticCalculator {
+public class ArithmeticCalculator{ // <T extends Number,S extends Number>{
 
     //연산 결과를 저장하는 컬렉션 타입 필드
     private List<Double> record = new ArrayList<>();
@@ -34,25 +16,25 @@ public class ArithmeticCalculator {
         this.record.add(recordValue);
     }
 
-    public Double calculate(int num1, int num2, OperatorType operator) {
+    //피연산자를 여러 타입으로 받을 수 있도록 기능을 확장
+    public <T extends Number, S extends Number> double calculate(T num1, S num2, OperatorType operator) {
 
-        double result = 0;
+       double result = 0;
 
         switch (operator) {
             case ADD:
-                result = num1 + num2;
+                result = num1.doubleValue() + num2.doubleValue();
                 break;
             case SUBTRACT:
-                result = num1 - num2;
+                result = num1.doubleValue() - num2.doubleValue();
                 break;
             case MULTIPLY:
-                result = num1 * num2;
+                result = num1.doubleValue() * num2.doubleValue();
                 break;
             case DIVIDE:
-                result = (double) num1 / num2;
+                result = num1.doubleValue() / num2.doubleValue();
                 break;
         }
-
         return result;
     }
